+++ visualization/app/codeCharta/state/store/appSettings/showMetricLabelNodeName/showMetricLabelNodeName.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/createSelector"
+import { appSettingsSelector } from "../appSettings.selector"
+
+export const showMetricLabelNodeNameSelector = createSelector([appSettingsSelector], appSettings => appSettings.showMetricLabelNodeName)