+++ visualization/app/codeCharta/ui/viewCube/centerMapButton/centerMapButton.module.ts
@@ -1,0 +1,9 @@
+import { NgModule } from "@angular/core"
+import { CenterMapButtonComponent } from "./centerMapButton.component"
+
+@NgModule({
+	declarations: [CenterMapButtonComponent],
+	exports: [CenterMapButtonComponent],
+	entryComponents: [CenterMapButtonComponent]
+})
+export class CenterMapButtonModule {}