+++ visualization/app/codeCharta/ui/viewCube/viewCube.component.ts
@@ -1,4 +1,3 @@
-import "./viewCube.component.scss"
 import {
 	AmbientLight,
 	AxesHelper,