+++ visualization/app/material/material.module.ts
@@ -15,7 +15,9 @@ import { MatFormFieldModule } from "@angular/material/form-field"
 import { MatInputModule } from "@angular/material/input"
 import { MatProgressSpinnerModule } from "@angular/material/progress-spinner"
 import { MatListModule } from "@angular/material/list"
+import { MatExpansionModule } from "@angular/material/expansion"
 import { MatSliderModule } from "@angular/material/slider"
+import { MatSlideToggleModule } from "@angular/material/slide-toggle"
 
 const materialModules = [
 	MatSelectModule,
@@ -34,7 +36,10 @@ const materialModules = [
 	MatInputModule,
 	MatProgressSpinnerModule,
 	MatListModule,
-	MatSliderModule
+	MatExpansionModule,
+	MatSliderModule,
+	MatSlideToggleModule,
+	BrowserAnimationsModule
 ]
 
 @NgModule({